Guarantee recent ontospy version
For reasons I can't figure out, a recent Github Actions build using this
code base causes Ontospy to be imported at version 1.6.5.6, which was
released on 2016-02-25.

This patch adds a requirement to use the latest release (1.9.8.3,
released 2019-07-28).  I expect this could cause build or runtime
errors, but I don't know what this repository's test procedure is.